requirements.txt
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.0rc
python-dotenv
twilio
